requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
blis==1.2.0
boto3==1.36.20
botocore==1.36.20
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
cymem==2.0.11
dask==2025.1.0
fastparquet==2024.11.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
locket==1.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.24.2
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.0.0
polars==1.21.0
preshed==3.0.9
psutil==6.1.1
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.2
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
swifter==1.4.0
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xgboost==2.1.4
zipp==3.21.0
